ppo continuous is still wrong

+++ dopamine/replay_memory/circular_replay_buffer.py
@@ -195,7 +195,7 @@ class OutOfGraphReplayBuffer(object):
     storage_elements = [
         ReplayElement('observation', self._observation_shape,
                       self._observation_dtype),
-        ReplayElement('action', (), np.int32),
+        ReplayElement('action', (), np.float32),
         ReplayElement('reward', (), np.float32),
         ReplayElement('terminal', (), np.uint8),
         ReplayElement('old_pro', (), np.float32)
@@ -518,7 +518,7 @@ class OutOfGraphReplayBuffer(object):
     transition_elements = [
         ReplayElement('state', (batch_size,) + self._state_shape,
                       self._observation_dtype),
-        ReplayElement('action', (batch_size,), np.int32),
+        ReplayElement('action', (batch_size,), np.float32),
         ReplayElement('reward', (batch_size,), np.float32),
         ReplayElement('next_state', (batch_size,) + self._state_shape,
                       self._observation_dtype),
